requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Scheduling
apscheduler==3.10.4

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# HTTP client enhancements
httpx==0.25.2
tenacity==8.2.3

# Data validation and serialization
marshmallow==3.20.1

# Testing (dev dependencies)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0
responses==0.24.1

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
pre-commit==3.6.0

# Security
cryptography==41.0.7

# CLI
click==8.1.7
rich==13.7.0

# Utilities
python-dateutil==2.8.2
tzlocal==5.2